+++ tests/scheduler/backends/graph_backends/test_qblox_backend.py
@@ -41,15 +41,12 @@ def test_compiles_standard_schedules(
     """
 
     config = compile_config_basic_transmon_qblox_hardware
-    assert config.name == "Qblox backend"
+    assert config.name == "Qblox compiler"
     assert (
         config.backend == "quantify_scheduler.backends.graph_compilation.SerialCompiler"
     )
 
-    backend = SerialCompiler("Qblox backend")
-
-    # assert that no exception is raised.
-    # Act
+    backend = SerialCompiler(config.name)
     comp_sched = backend.compile(schedule=schedule, config=config)
     # Assert that no exception was raised and output is the right type.
     assert isinstance(comp_sched, CompiledSchedule)