+++ tests/scheduler/backends/test_graph_compilation.py
@@ -6,19 +6,15 @@ file.
 # pylint: disable=missing-module-docstring
 # pylint: disable=missing-class-docstring
 # pylint: disable=missing-function-docstring
+
 import pytest
 import networkx as nx
 from matplotlib.axes import Axes
-from quantify_scheduler.backends.graph_compilation import (
-    QuantifyCompiler,
-    SimpleNode,
-)
+
 from quantify_scheduler import Schedule
+from quantify_scheduler.backends.graph_compilation import QuantifyCompiler, SimpleNode
+from quantify_scheduler.operations.gate_library import Reset
 
-
-from quantify_scheduler.operations.gate_library import (
-    Reset,
-)
 
 # pylint: disable=unused-argument
 def dummy_compile_add_reset_q0(schedule: Schedule, config=None) -> Schedule: